pyproject.toml
[tool.poetry]
name = "shutdown timer"
version = "0.1.0"
description = ""
authors = ["elsheikh21 <ahmedhelsheikh@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
black = "^21.10b0"
PyQt5 = "^5.15.6"
PyYaml = "^6.0"
pyinstaller = "^4.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
